+++ app/src/main/java/com/jorgecastanov/mercadolibrecandidate/data/api/ProductApi.kt
@@ -6,7 +6,7 @@ import retrofit2.http.Query
 
 private const val SITE_ID = "MLA"
 
-interface ApiService {
+interface ProductApi {
 
     @GET("/sites/$SITE_ID/search")
     suspend fun getProducts(
