--- app/src/main/java/com/jorgecastanov/mercadolibrecandidate/data/repository/FeedRepositoryImpl.kt
@@ -1,14 +1,0 @@
-package com.jorgecastanov.mercadolibrecandidate.data.repository
-
-import com.jorgecastanov.mercadolibrecandidate.data.api.ApiHelper
-import com.jorgecastanov.mercadolibrecandidate.data.model.Product
-import javax.inject.Inject
-
-class FeedRepositoryImpl @Inject constructor(
-    private val apiHelper: ApiHelper
-    ) : FeedRepository {
-
-    override suspend fun getProducts(keyWord: String): List<Product> =
-        apiHelper.getProducts(keyWord)
-
-}
