+++ app/src/main/java/com/jorgecastanov/mercadolibrecandidate/data/repository/ProductRepository.kt
@@ -2,7 +2,7 @@ package com.jorgecastanov.mercadolibrecandidate.data.repository
 
 import com.jorgecastanov.mercadolibrecandidate.data.model.Product
 
-interface FeedRepository {
+interface ProductRepository {
 
     suspend fun getProducts(keyWord: String): List<Product>
