+++ app/src/main/java/com/jorgecastanov/mercadolibrecandidate/di/RepositoryModule.kt
@@ -1,8 +1,8 @@
 package com.jorgecastanov.mercadolibrecandidate.di
 
-import com.jorgecastanov.mercadolibrecandidate.data.api.ApiHelper
-import com.jorgecastanov.mercadolibrecandidate.data.repository.FeedRepository
-import com.jorgecastanov.mercadolibrecandidate.data.repository.FeedRepositoryImpl
+import com.jorgecastanov.mercadolibrecandidate.data.datasource.ProductDataSource
+import com.jorgecastanov.mercadolibrecandidate.data.repository.ProductRepository
+import com.jorgecastanov.mercadolibrecandidate.data.repository.ProductRepositoryImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -15,7 +15,7 @@ class RepositoryModule {
 
     @Singleton
     @Provides
-    fun provideFeedRepository(apiHelper: ApiHelper): FeedRepository =
-        FeedRepositoryImpl(apiHelper)
+    fun provideProductRepository(productDataSource: ProductDataSource): ProductRepository =
+        ProductRepositoryImpl(productDataSource)
 
 }
