+++ app/src/main/java/com/jorgecastanov/mercadolibrecandidate/ui/screens/feed/FeedState.kt
